remove usesOpenGLAcceleration method and it's single use in QSAboutWindowController most (unsigned) ints are NS(U)Integers some floats are now CGFloats some methods in NDClasses, mostly NDAppleScript related stuff, have been commented out for now. they should probably be removed rather than updated build settings for 64-bit

+++ Quicksilver/Code-App/QSTriggersPrefPane.h
@@ -52,7 +52,7 @@
 
 	QSTrigger *selectedTrigger;
 	NSSplitView *splitView;
-	int selectedRow;
+	NSInteger selectedRow;
 }
 - (NSString *)currentSet;
 - (void)setCurrentSet:(NSString *)value;
@@ -81,8 +81,8 @@
 
 - (IBAction)showTriggerInfo:(id)sender;
 - (IBAction)hideTriggerInfo:(id)sender;
-- (int)tabViewIndex;
-- (void)setTabViewIndex:(int)index;
+- (NSInteger)tabViewIndex;
+- (void)setTabViewIndex:(NSInteger)index;
 
 - (IBAction)editCommand:(id)sender;
