+++ Quicksilver/Code-QuickStepFoundation/NSApplication_BLTRExtensions.m
@@ -82,7 +82,7 @@
 	[self relaunchAtPath:[[NSBundle mainBundle] bundlePath] movedFromPath:newPath];
 }
 
-- (int)moveToPath:(NSString *)launchPath fromPath:(NSString *)newPath {
+- (NSInteger)moveToPath:(NSString *)launchPath fromPath:(NSString *)newPath {
 	NSFileManager *manager = [NSFileManager defaultManager];
 	NSString *tempPath = [[launchPath stringByDeletingLastPathComponent] stringByAppendingPathComponent:[NSString stringWithFormat:@"%@.old.app",[[NSProcessInfo processInfo] processName]]];
 	//NSLog(@"temp %@ new %@", tempPath, newPath);
@@ -152,7 +152,7 @@
 @end
 
 @implementation NSApplication (LaunchStatus)
-- (int)checkLaunchStatus {
+- (NSInteger)checkLaunchStatus {
 	NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
 	NSString *lastLocation = [defaults objectForKey:kLastUsedLocation];
 
@@ -162,8 +162,8 @@
 	if (! (lastLocation || lastVersionString) )
 		return QSApplicationFirstLaunch;
 
-	int lastVersion = [lastVersionString respondsToSelector:@selector(hexIntValue)] ? [lastVersionString hexIntValue] : 0;
-	int thisVersion = [thisVersionString hexIntValue];
+	NSInteger lastVersion = [lastVersionString respondsToSelector:@selector(hexIntValue)] ? [lastVersionString hexIntValue] : 0;
+	NSInteger thisVersion = [thisVersionString hexIntValue];
 
 	if (thisVersion>lastVersion) return QSApplicationUpgradedLaunch;
 	if (thisVersion<lastVersion) return QSApplicationDowngradedLaunch;
